requirements.txt
PySide6>=6.9.0
faster-whisper>=1.1.0
torch>=2.0.0
PyYAML>=6.0
psutil>=5.9.0
tqdm>=4.65.0

